Use Gradle categories to mark headless tests
